Add functionality for Chess' checking rule
The rule is implemented in IsLegalMove by constructing a board
representing the boardstate after the move has been made. We pass
this board to the IsInCheck function. If IsInCheck returns true then
IsLegalMove returns false.

The IsInCheck function calls IsLegalMove n times (where n equals the
number of the opponents pieces) with an opponent's piece as the origin
and the friendly king as the target. If any of these calls returns true
then IsInCheck return true.

Note: IsLegalMove takes considerChecks as an argument because a pinned
piece can still give check: i.e. a piece that is pinned against the king
can still move to kill the enemy king even if doing so leaves its own
king in check.