Decouple board state from ChessTiles
A new class called ChessBoard stores the posistions of pieces in 0x88
form. ChessBoard has accessors that allow the caller to get or set the
piece at a given board position either by a single integer position or
by rank-file.

ChessTile now has a new constructor that associates the tile with its
ChessBoard. This allows us to change the getters and setters for
ChessTile.PieceType to essentially reference the ChessBoard state.