Big board/tile interaction refactor
- Tidied up old constructors
- Initialise more things in tile constructors
- Remove SetPiece, replaced by board[pos]
- Refactor ParseFen to return ChessPiece[] instead of ChessTile[]
- Give ChessTile an Update() method that sets the tile's bitmap

Both BoardViewModel and ChessBoard now have references to each other so
that each BoardViewModel knows which board it's associated with and
each ChessBoard knows which BoardViewModel it's associated with. This
allows us to access the ChessTiles from the ChessBoard class.
Additionally, every tile now stores a reference to the board that it is
on which allows us to know which piece is on any given tile.

Whenever the board state is changed through board[pos], the Update()
method of the relevant ChessTile is called so that the bitmaps are
updated properly. This seems easier to understand than an event driven
system for the moment since if we want to avoid calling Update on every
single tile we need to create a new EventArgs class that stores the
information about which tile was changed. We can always move to an event
driven system in the future but having an Update() method seems to make
the most sense at the moment.