Spawn the engine as a child process
We now have a cmake wrapper for building this project. You can use cmake
as you usually would and it will also invoke the dotnet commands to
build the GUI.

NOTE: You need to run the project with dotnet build\chess.dll instead of
dotnet run -p chess, otherwise the gui will not find the engine.

closes #19