Fix ChessTile colour code, ChessTimer, promotions
If the king is moved, then put in check on the next ply, the king's tile
does not turn red. This is because the tiles colour is cleared when
tiles_to_clear is cleared (in UpdateMoveFill). To fix this, I made it so
that the tiles colour is calculated in the Fill getter using multiple
bools. To change the Fill, change one of the bools.

Timer doesn't update when it runs out

Pieces can be promoted when IsInteractable = false.

Pieces can be moved when IsPromoting = true.

Promotion didn't work well with move history navigation. Promoted pawns
are displayed as pawns instead of the promoted piece (for the move at
which the promotion occured) when clicking "Prev" and "Next". Closes #48.

After promoting, no checks are done for IsInCheck or IsInStalemate.