Use Dictionary for GUID matching
This means that getting the corresponding reply for a request is just a
dictionary lookup.

NOTE: We now REUSE the same Message object for both sending a receiving.
This is so that the GUIDs match. Note that we can overwrite whatever
data is in mess.data in mess.Receive since if we're receiving the
message, we must have already sent it which means the data in mess.data
has already been delivered.