Message rework (#50)
* Add some extension methods

* Read/write messages into lists

We now have threads that constantly read/write messages to and from
lists. Almost like a message queue, but not quite. This will allow us to
avoid some syncronisation problems like not having to immediately reply
with a message all the time.

* Make everything shutdown on window close

If we don't do this then we end up with some dangling treads and the
engine is not terminated properly

* Use Dictionary for GUID matching

This means that getting the corresponding reply for a request is just a
dictionary lookup.

NOTE: We now REUSE the same Message object for both sending a receiving.
This is so that the GUIDs match. Note that we can overwrite whatever
data is in mess.data in mess.Receive since if we're receiving the
message, we must have already sent it which means the data in mess.data
has already been delivered.

* fixup! Use Dictionary for GUID matching

* Refactor loop

* Send the header in one call